+++ ReleaseNotes.md
@@ -9,10 +9,11 @@ ELF-specific artifacts have been moved to https://github.com/elf-oskari/oskari-s
 
 ### DB upgrades
 
-#### Allowing more data to be stored for users longer usernames
+#### Allowing more data to be stored for users
 
 * Usernames in external systems can be longer than current column size - adjusted to 128 characters.
 * User UUID should be unique - added constraint.
+* Added column for users email.
 * User can now store additional attribute data as JSON.
 
     content-resources/src/main/resources/sql/upgrade/1.27/03_alter_table_oskari_users.sql
@@ -22,12 +23,15 @@ ELF-specific artifacts have been moved to https://github.com/elf-oskari/oskari-s
 DatabaseUserService now has a method (saveUser) to ensure the user and his/her roles are up-to-date in the database (inserting if missing).
 Usable when combined with external login service which should populate users to Oskari based on external user data (for example SAML-message).
 Added support for saving and loading additional user attributes to/from database.
+User email is now saved to the database.
 
 ### service-base
 
 IOHelper now has convenience method to write the Content-type header.
 HTTP params can now be generated to properly encoded string for encoding POST payload or GET querystring with IOHelper.getParams().
 User class now has methods to attach additional attributes for the user.
+PropertyUtil.getNecessary() has a new overloaded version that takes an optional detailed message. The message is attached to the
+exception so it is shown in the server logs. Message should be used to tell why the property is necessary.
 
 ### geoserver-ext/OskariMarkFactory